Use POST for /run/delete and add confirmation form

+++ src/Xhgui/Controller/Run.php
@@ -108,20 +108,42 @@ class Xhgui_Controller_Run extends Xhgui_Controller
         ));
     }
 
-    public function delete()
+    public function deleteForm()
     {
         $request = $this->app->request();
         $id = $request->get('id');
+        if (!is_string($id) || !strlen($id)) {
+            throw new Exception('The "id" parameter is required.');
+        }
+
+        // Get details
+        $result = $this->profiles->get($id);
+
+        $this->_template = 'runs/delete-form.twig';
+        $this->set(array(
+            'run_id' => $id,
+            'result' => $result,
+        ));
+    }
+
+    public function deleteSubmit()
+    {
+        $request = $this->app->request();
+        $id = $request->post('id');
+        // Don't call profilers->delete() unless $id is set,
+        // otherwise it will turn the null into a MongoId and return "Sucessful".
+        if (!is_string($id) || !strlen($id)) {
+            // Form checks this already,
+            // only reachable by handcrafted or malformed requests.
+            throw new Exception('The "id" parameter is required.');
+        }
 
         // Delete the profile run.
         $delete = $this->profiles->delete($id);
 
         $this->app->flash('success', 'Deleted profile ' . $id);
 
-        $referrer = $request->getReferrer();
-        // In case route is accessed directly the referrer is not set.
-        $redirect = isset($referrer) ? $referrer : $this->app->urlFor('home');
-        $this->app->redirect($redirect);
+        $this->app->redirect($this->app->urlFor('home'));
     }
 
     public function deleteAllForm()
@@ -138,10 +160,7 @@ class Xhgui_Controller_Run extends Xhgui_Controller
 
         $this->app->flash('success', 'Deleted all profiles');
 
-        $referrer = $request->getReferrer();
-        // In case route is accessed directly the referrer is not set.
-        $redirect = isset($referrer) ? $referrer : $this->app->urlFor('home');
-        $this->app->redirect($redirect);
+        $this->app->redirect($this->app->urlFor('home'));
     }
 
     public function url()
